+++ src/HasTranslations.php
@@ -36,12 +36,13 @@ trait HasTranslations
     /***
      * @param string $key
      * @param string $locale
+     * @param bool $useFallbackLocale
      *
      * @return mixed
      */
-    public function getTranslation(string $key, string $locale)
+    public function getTranslation(string $key, string $locale, bool $useFallbackLocale = true)
     {
-        $locale = $this->normalizeLocale($key, $locale);
+        $locale = $this->normalizeLocale($key, $locale, $useFallbackLocale);
 
         $translations = $this->getTranslations($key);
 
@@ -52,6 +53,16 @@ trait HasTranslations
         }
 
         return $translation;
+    }
+
+    public function getTranslationWithFallback(string $key, string $locale)
+    {
+        return $this->getTranslation($key, $locale, true);
+    }
+
+    public function getTranslationWithoutFallback(string $key, string $locale)
+    {
+        return $this->getTranslation($key, $locale, false);
     }
 
     public function getTranslations($key) : array
@@ -141,9 +152,13 @@ trait HasTranslations
         }
     }
 
-    protected function normalizeLocale(string $key, string $locale) : string
+    protected function normalizeLocale(string $key, string $locale, bool $useFallbackLocale) : string
     {
         if (in_array($locale, $this->getTranslatedLocales($key))) {
+            return $locale;
+        }
+
+        if (!$useFallbackLocale) {
             return $locale;
         }