+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All Notable changes to `laravel-translatable` will be documented in this file
 
+## 1.2.2 - 2016-01-27
+- improve support for fallback locale
+
+## 1.2.1 - 2016-01-23
+- improve compatibility for Laravel 5.4
+
 ## 1.2.0 - 2016-01-23
 - add compatibility for Laravel 5.4